chore(warnings): nettoyage warnings build (~125 traités, dont 1 vrai bug) (#969)
## Nettoyage des warnings de build

Suite au constat de nombreux warnings dans le build. Traite **~125 des
139** warnings.

### Corrections de fond (correctness / sécurité)
- **Troncatures `size_t → uint32_t` (`C4267`, perte de données
possible)** — la préoccupation soulevée : `DecalPass`
(`binding`/`dstBinding`) + `TradeSystem` (`RemoveSession`) →
`static_cast<uint32_t>` explicite (indices petits, sûrs).
- **🐛 VRAI BUG (`C4125`)** — la liste de résolutions ImGui
`"1280x720\01600x900\0…"` : un `\0` suivi d'un chiffre était interprété
comme **échappement octal** (`\016` = caractère 14), **corrompant les
séparateurs** de la combo (résolutions mal découpées). Corrigé en
littéraux juxtaposés (`"…\0" "…\0"`), chaque `\0` étant alors un vrai
NUL.
- **Noms locaux masqués (`C4456`)** — `TerrainRenderer`
(`vbOffset`→`cliffVbOffset`), `ChatImGuiRenderer`
(`flags`→`inputFlags`).

### Réduction de bruit
- **En-têtes tiers (~75 warnings)** — `vk_mem_alloc.h`, `cgltf.h`,
`miniaudio.h` enveloppés dans `#pragma warning(push, 0)/pop` (gardé
`_MSC_VER`), **sans éditer le code vendored**.
- **Params/variables inutilisés (`C4100`/`C4189`, ~22)** — attribut
`[[maybe_unused]]` (non destructif, aucun changement de comportement).
- **`C4127` dans les tests de layout** (`REQUIRE` sur
`sizeof`/`offsetof` = expressions constantes) — `#pragma
warning(disable:4127)` en tête des fichiers de test concernés.

### Volontairement NON traité (chantier séparé)
- **`C4996` (~13)** — API dépréciées (OpenSSL `SHA256_*`, CRT).
**Migration prudente requise** (crypto/CRT sensibles) — à faire dans une
PR dédiée, pas en aveugle.
- 2 `C4127` hors tests + 1 `C4566` — mineurs.

### Comportement / déploiement
Aucun changement de comportement (les casts formalisent une conversion
déjà implicite ; les attributs/pragmas n'affectent pas la génération de
code). Fichiers serveur touchés (`ServerApp`/`TradeSystem`) mais
**neutres**.
> **Déploiement** : ✅ neutre — pas de redéploiement serveur fonctionnel
requis (aucun changement de protocole, gameplay ou schéma).

+++ src/client/render/DecalPass.cpp
@@ -83,7 +83,7 @@ namespace engine::render
 			std::array<VkDescriptorSetLayoutBinding, 2> bindings{};
 			for (size_t i = 0; i < bindings.size(); ++i)
 			{
-				bindings[i].binding = i;
+				bindings[i].binding = static_cast<uint32_t>(i);
 				bindings[i].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
 				bindings[i].descriptorCount = 1;
 				bindings[i].stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
@@ -385,7 +385,7 @@ namespace engine::render
 			{
 				writes[i].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
 				writes[i].dstSet = ds;
-				writes[i].dstBinding = i;
+				writes[i].dstBinding = static_cast<uint32_t>(i);
 				writes[i].descriptorCount = 1;
 				writes[i].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
 				writes[i].pImageInfo = &imageInfos[i];
